部分编译错误以及警告更改

+++ src/fsmc_nor.h
@@ -16,17 +16,17 @@
 #ifndef __FSMC_NOR_H
 #define __FSMC_NOR_H
 
-#define Bank1_NOR2_ADDR       ((u32)0x64000000)
+#define Bank1_NOR2_ADDR       ((long)0x64000000)
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f10x_lib.h"
 
 /* Exported types ------------------------------------------------------------*/
 typedef struct {
-	u16 Manufacturer_Code;
-	u16 Device_Code1;
-	u16 Device_Code2;
-	u16 Device_Code3;
+	short Manufacturer_Code;
+	short Device_Code1;
+	short Device_Code2;
+	short Device_Code3;
 } NOR_IDTypeDef;
 
 /* NOR Status */
@@ -42,17 +42,17 @@ typedef enum {
 /* Exported functions ------------------------------------------------------- */
 void NOR_Init(void);
 void FSMC_NOR_ReadID(NOR_IDTypeDef *NOR_ID);
-NOR_Status FSMC_NOR_EraseSector(u32 BlockAddr);
-NOR_Status FSMC_NOR_EraseBlock(u32 BlockAddr);
+NOR_Status FSMC_NOR_EraseSector(long BlockAddr);
+NOR_Status FSMC_NOR_EraseBlock(long BlockAddr);
 NOR_Status FSMC_NOR_EraseChip(void);
-NOR_Status FSMC_NOR_WriteHalfWord(u32 WriteAddr, u16 Data);
-NOR_Status FSMC_NOR_WriteBuffer(u16 *pBuffer, u32 WriteAddr, u32 NumHalfwordToWrite);
-NOR_Status FSMC_NOR_ProgramBuffer(u16 *pBuffer, u32 WriteAddr, u32 NumHalfwordToWrite);
-u16 FSMC_NOR_ReadHalfWord(u32 ReadAddr);
-void FSMC_NOR_ReadBuffer(u16 *pBuffer, u32 ReadAddr, u32 NumHalfwordToRead);
+NOR_Status FSMC_NOR_WriteHalfWord(long WriteAddr, short Data);
+NOR_Status FSMC_NOR_WriteBuffer(short *pBuffer, long WriteAddr, long NumHalfwordToWrite);
+NOR_Status FSMC_NOR_ProgramBuffer(short *pBuffer, long WriteAddr, long NumHalfwordToWrite);
+short FSMC_NOR_ReadHalfWord(long ReadAddr);
+void FSMC_NOR_ReadBuffer(short *pBuffer, long ReadAddr, long NumHalfwordToRead);
 NOR_Status FSMC_NOR_ReturnToReadMode(void);
 NOR_Status FSMC_NOR_Reset(void);
-NOR_Status FSMC_NOR_GetStatus(u32 Timeout);
+NOR_Status FSMC_NOR_GetStatus(long Timeout);
 
 #endif /* __FSMC_NOR_H */
